plots/requirements.txt
asyncio
asyncpg
python-dotenv
wheel
argparse
matplotlib
astropy==5.2.0
astroquery
numpy
retrying-async
